코드 정리 및 Old Editor 삭제
