+++ src/App.tsx
@@ -1,12 +1,11 @@
 import Layout from "./components/Layout";
 import Editor from "./components/Editor";
-import NewEditor from "./components/NewEditor";
 
 const App = () => {
   return (
     <div className="App">
       <Layout>
-        <NewEditor />
+        <Editor />
       </Layout>
     </div>
   );
